src/etl/utils.py
import yaml

def load_yaml(path: str):
    try:
        with open(path, "r") as file:
            return yaml.safe_load(file)
    except yaml.YAMLError as e:
        raise ValueError(f"Error loading YAML file '{path}': {str(e)}") from e