requirements.txt
astroid==3.3.10
cfgv==3.4.0
dill==0.4.0
distlib==0.3.9
filelock==3.18.0
identify==2.6.12
isort==6.0.1
-e git+ssh://git@github.com/Karmendes/karmendes_etl.git@178856297722aec5d2112565f93a864277b993dd#egg=karmendes_etl
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.3.8
pre_commit==4.2.0
pylint==3.3.7
PyYAML==6.0.2
tomlkit==0.13.3
virtualenv==20.31.2
